requirements.txt
PyQt6
yt_dlp
whisper
openai
langchain
langchain-openai
python-dotenv
ffmpeg-python